Move remaining common code from spdx-utils to common-utils
