+++ utils/core/src/main/kotlin/ArchiveUtils.kt
@@ -49,9 +49,9 @@ import org.apache.commons.compress.compressors.xz.XZCompressorInputStream
 import org.apache.commons.compress.utils.SeekableInMemoryByteChannel
 
 import org.ossreviewtoolkit.utils.common.Os
+import org.ossreviewtoolkit.utils.common.VCS_DIRECTORIES
 import org.ossreviewtoolkit.utils.common.safeDeleteRecursively
 import org.ossreviewtoolkit.utils.common.safeMkdirs
-import org.ossreviewtoolkit.utils.spdx.VCS_DIRECTORIES
 
 enum class ArchiveType(vararg val extensions: String) {
     TAR(".gem", ".tar"),
